Добавил защиту от удаления Default пользователя

+++ StarkCNC/Services/UserService.cs
@@ -131,6 +131,9 @@ public class UserService : IUserService
         if (!CanUserUpdate(user, RolePermissions.IdentityRoleToRoles(userRole?.Name)))
             throw new InvalidOperationException("Ваш уровень доступа не позволяет вам удалять пользователей данного уровня доступа");
 
+        if (user.UserName == "Default")
+            throw new InvalidOperationException("Нельзя удалить пользователя по умолчанию");
+
         var userManager = _serviceProvider.GetRequiredService<UserManager<User>>();
 
         await userManager.DeleteAsync(user).ConfigureAwait(false);
